requirements.txt
fastapi==0.68.1
motor==2.5.1
python-dotenv
httpx
uvicorn